added testing nodejs to sql
