+++ client/home.html
@@ -47,11 +47,11 @@
                 </tr>
                 <tr>
                     <th>User:</th>
-                    <td><input type="text" name="createAcc" id="user"></td>
+                    <td><input type="text" name="login" id="user"></td>
                 </tr>
                 <tr>
                     <th>Password:</th>
-                    <td><input type="text" name="createAcc" id="pass"></td>
+                    <td><input type="text" name="login" id="pass"></td>
                 </tr>
                 <tr>
                     <td><input type="submit" name="submit" value="Login"></button></td>
