+++ client/view_teams.html
